Moves to routes dir
